fix bug caused by replacing fragment with Scene
Cannot use react-scrollmagics' Scene component like a fragment
to create conditionally rendered components. Doing so causes a
"React.Children.only expected to receive a single React element
child" error to occur because Scene returns the children given to it.

This fix re-adds the fragment around the component and nests the Scene
component inside the conditional render expression for product link list
items — guaranteeing that a single item is always returned.